fix(rust): 修复 7 个编译错误 + 启动配置
- output_parser.rs: 修复 powi 在 ambiguous float 上的调用
- md_postprocess.rs: 修复 f64 % integer 类型不匹配
- phase_field.rs: 补全 PfConfig 缺失的 output_interval 字段
- cantilever_beam_test.rs: 修复 caelab_lib → caelab、补全缺失括号
- tauri.conf.json: 移除不支持的 ios bundle 配置
- db/mod.rs: 初始化 knowledge_memory + vector_store 表
- ml_potential.rs: 补充 mock 训练注释
- 新增 learning_tracker.rs

+++ src-tauri/src/commands/md_postprocess.rs
@@ -938,8 +938,8 @@ mod tests {
                 let t = frame as f64 * 0.1;
                 // Random walk-like displacement
                 frame_pos.push([
-                    atom as f64 * 3.0 + t * 0.5 * (atom as f64 % 3 + 1) as f64,
-                    atom as f64 * 3.0 + t * 0.3 * (atom as f64 % 2 + 1) as f64,
+                    atom as f64 * 3.0 + t * 0.5 * (atom as f64 % 3.0 + 1.0),
+                    atom as f64 * 3.0 + t * 0.3 * (atom as f64 % 2.0 + 1.0),
                     atom as f64 * 3.0 + t * 0.4,
                 ]);
             }
